+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.5] - 2022-12-22
+### Fixed
+- Fix an incorrect zooming behaviour on ScreenSpace-Camera render mode.
+### Updated
+- Optimize Calculation during zoom deceleration.
+- Update a proper way to handle stretched anchor content.
+
 ## [1.0.4] - 2022-09-15
 ### Fixed
 - Fix an incorrect zooming behaviour on ScreenSpace-Camera & WorldSpace render mode.